Location Encoding

+++ src/main/java/me/outspending/position/Location.java
@@ -1,6 +1,6 @@
 package me.outspending.position;
 
-public record Location(double x, double y, double z) {
+public record Location(long x, long y, long z) {
     public static Location ZERO = new Location(0, 0, 0);
     public static Location ONE = new Location(1, 1, 1);
 }
